Added PSR-4 autoloading, separated into component files.
