+++ src/Card.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Forthplace\Candyland;
+
+class Card {
+  public $type;
+  public $double;
+
+  public function __construct($type, $double = FALSE) {
+    $this->type = $type;
+    $this->double = $double;
+  }
+}
